scraper/requirements.txt
anyio==3.6.2
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
decorator==5.1.1
exceptiongroup==1.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
mypy==0.991
mypy-extensions==0.4.3
packaging==23.0
pika==1.3.1
pluggy==1.0.0
pycurl==7.45.2
pytest==7.2.1
python-dateutil==2.8.2
requests==2.28.2
retry==0.9.2
rfc3986==1.5.0
ruff==0.0.237
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tomli==2.0.1
types-beautifulsoup4==4.11.6.7
types-html5lib==1.1.11.12
types-pika==1.2.0b1
types-pycurl==7.45.2.3
types-python-dateutil==2.8.19.8
types-requests==2.28.11.14
types-retry==0.9.9.2
types-urllib3==1.26.25.7
typing_extensions==4.4.0
urllib3==1.26.14
